Add new utils file for kin var additions
